Crate keyboard_types

Contains types to define keyboard related events.

The naming and conventions follow the UI Events specification but this crate should be useful for anyone implementing keyboard input in a cross-platform way.

Modules

code 🔒

composition 🔒

key 🔒

key_state 🔒

keyboard_event 🔒

location 🔒

modifiers 🔒

Modifier key data.

named_key 🔒

shortcuts 🔒

webdriver

Keyboard related WebDriver functionality.

Structs

CompositionEvent

Event to expose input methods to program logic.

KeyboardEvent

Keyboard events are issued for all pressed and released keys.

Modifiers

Pressed modifier keys.

ShortcutMatcher

Match keyboard shortcuts and execute actions.

UnrecognizedCodeError

Parse from string error, returned when string does not match to any Code variant.

UnrecognizedKeyError

Parse from string error, returned when string does not match to any Key variant.

UnrecognizedNamedKeyError

Parse from string error, returned when string does not match to any NamedKey variant.

Enums

Code

Code is the physical position of a key.

CompositionState

Describes the state of a composition session.

Key

The value received from the keypress.

KeyState

Describes the state a key is in.

Location

The location attribute contains an indication of the physical location of the key on the device.

NamedKey

Key represents the meaning of a keypress.

Functions

first_char 🔒

Return the first codepoint of a string.