Crate keyboard_types

Contains types to define keyboard related events.

The naming and conventions follow the UI Events specification but this crate should be useful for anyone implementing keyboard input in a cross-platform way.

Modules

code

key

location

modifiers
Modifier key data.

shortcuts

webdriver
Keyboard related WebDriver functionality.

Structs

CompositionEvent
Event to expose input methods to program logic.

KeyboardEvent
Keyboard events are issued for all pressed and released keys.

Modifiers
Pressed modifier keys.

ShortcutMatcher
Match keyboard shortcuts and excute actions.

Enums

Code
Code is the physical position of a key.

CompositionState
Describes the state of a composition session.

Key
Key represents the meaning of a keypress.

KeyState
Describes the state the key is in.

Location
The location attribute contains an indication of the logical location of the key on the device.

Functions

first_char
Return the first codepoint of a string.